Commit message: Refactor IdeaPipeline and create Filter base class
The commit refactors the IdeaPipeline by creating a base class, Filter, for filters that can be used in a pipeline. It also updates the OpenAIAssistant class to inherit from Filter, and removes the num_requests parameter from its constructor. Instead, this parameter is now passed in as a keyword argument when the class is initialized.

The commit also updates the IdeaGenerator and IdeaRanker classes to use the new OpenAIAssistant constructor, and changes their constructor arguments to keyword arguments.

Finally, the commit updates the Main.py file to use the new Filter and OpenAIAssistant classes, and removes the num_requests parameter from the IdeaGenerator and IdeaRanker constructors.